- Introducing a new network and command queue ready HandshakeData type for handling handshakes, alongside other minor changes
- Removed probe event flags and reused NetLinkState directly.
- Instance registry now transmits the actual connection state and
introducting new helper functions to retrieve connection and session
states.
- Changed monitor res internal data type to uint32 to be explicit.

+++ src/misc/platform/windows/system_probe.cpp
@@ -11,7 +11,8 @@
 
 Device::MonitorRes Device::GetMonitorResolution()
 {
-    return MonitorRes{GetSystemMetrics(SM_CXSCREEN), GetSystemMetrics(SM_CYSCREEN)};
+    return MonitorRes{static_cast<uint32_t>(GetSystemMetrics(SM_CXSCREEN)),
+                      static_cast<uint32_t>(GetSystemMetrics(SM_CYSCREEN))};
 }
 
 void Device::RetrieveUserName(char (&CharArray)[MAX_UNLEN])
